fix: Alpaca getBars 최신 데이터 반환하도록 수정
